requirements.txt
click
fusepy
python-logstash-async
psutil
